Remove the Nova aggregate files.
There was a commit [0] that dealt with Nova aggs that also updated
doc/source/user/placement.rst. Since that file was preserved in the
filter_history.sh script, that commit caused a lot of agg-related files
to be preserved. This commit removes most of them; others in the test
directories will be removed in a later patch.

[0] https://review.openstack.org/#/c/553597/

Change-Id: I2b0ee94c1d7f11df542ec5171278696fb21b11d1