Set up initial .zuul.yaml
Set up placement with a noop job for 'gate' and 'check' and non-voting
unit, functional and pep8 test for 'check'. These jobs will be change
to non-voting as they reach a working state, and additional jobs will
be added as required.

The noop jobs are required for any code to merge.

Also .gitreview is updated to point to the right repo.

Change-Id: I741a6b86eda12afb515b6207a4535e8b6e1ead44