Inspect and correct tox.ini, .stestr.conf and setup.cfg
Begin the process of getting placement tests running, by getting the
test harness opertional.

Before we can expect the tests to be able to run the configuration for
the placement package and its test harness needs to be adjusted for
"nova" -> "placement". These changes are made by inspection and based
on previous iterations of doing these kinds of cleanups, it won't be
until later in this stack of change that actually passing tests will
show up. However, the logs for attempted tests runs should be getting
a little bit further than the parent.

setup.cfg changes quite a bit because there are many things in it that
are not relevant. placement (thus far) really only has one entry point.
As noted, by a TODO, a placement-manage and placement-status will be
desired in the near future.

Change-Id: Iaa385c6e8653a9646a21bb65203deaa05ec43217