Update i18n handling to be placement oriented.
The translation domain is set to 'placement' and log translation is
removed. Imports are adjusts to reflect the new location.

placement/utils.py to only import _, and none of the log translation
methods. These are only used in code that will be removed further
along in this sequence of changes.

Change-Id: Icf9bae4e2f0650b5eb4697afdc287de10153ce24