Remove some imports from test/fixtures that will not be used
This doesn't get all of the unused imports, nor does it remove the
code associated with the removed imports. This just removes a suite
of imports that are blocking further progress on attempting to get
functional tests working. Later changes will return to this file
to remove more imports and to remove dead code.

NOTE: The 'from  import wsgi_app' line is a leftover from mechanical
processing of path renames earlier in the extraction process.

Change-Id: I41f943d73b3c23523d265cf508e4d29b2939aa13