Establish an importable 'conf' package
This sets up the basic configuration needs for placement and removes
extraneous imports for files that were not included as part of the
extraction. This does not yet actually test the configuration, just
makes it so that it can be imported and eventually tested further along
in this sequence of commits.

Change-Id: I4d4f378dbaf2cde70eb9adecc2b9d87814097a23