Import placement, not nova, in rp db tests
Within the tests some mocks are done by package, not string, so
the import is required to get those tests running properly. Which this
change the following is successful:

.tox/py27/bin/python -m testtools.run discover -f placement.tests.functional.db
Tests running...
[sqlachemy warning noise deleted]
Ran 138 tests in 6.095s
OK

This success will not show up in zuul-based tests yet, as stestr's discovery
mechanism is not yet able to complete.

Change-Id: If8c1fa52ecc3b8afbcc33749827cb082facf4aac