Trim placement/utils.py to the single method used
The tempdir method is used by the test_policy unit tests, which
will be made to work further along in this sequence.

Later we may wish to either move the tempdir method somewhere else
or collapse utils.py and util.py into one file. Not doing that now
as these changes are devoted to get the tests passing with minimal
changes.

Change-Id: Ic2d67d30ab13f43a52ef5d1dfd78b9dc4fe4024e