Remove unused fixtures in placement/tests/fixtures.py
This gets pep8 happy across the board.

In the near future the fixtures that are scattered around in the
placement repo should be consolidated and/or moved to more sensible
locations. The database fixture (the one remaining in this file) is
only needed by functional tests so should move to
placement/tests/functional/fixutures. That is not being done as part
of this series of changes because this series is solely about
getting the tests passing. Other types of cleanup will come once the
tests are voting.

Change-Id: I8b657878b496e3d3c4d65e41509aefbcd6e396bf