Update requirements and test-requirements
With placement off by itself, the number of python packages required to
make things go shrinks.

This is the set of modules that are imported directly. There are plenty
of transitive dependencies not listed here.

test-requirements.txt also lists some of the optional things (database
drivers) that come up in devstack-based testing situations.

Change-Id: I3eb6aead544d1ef647ad8d94a18382d5fa795aa4