de-novify wsgi application to expect placement config
Replace references of 'nova' to 'placement' so that the wsgi
application loads up the right config and uses the right conf
options. We don't currently have tests for this (because we have
no dsvm-associated jobs) but this has been successfully used in
a hacked up local devstack.

This patch will be included as part of a Depends-On stack that
tests those devstack hacks.

Change-Id: Ibed4c64fa4e74c830d0d2fa040089bf9dc7b8f29