Remove multiple database scaffolding
Placement uses a single database, so the code required to support
multiple databases that came from Nova is not needed.

I also changed the global 'session_configured' to lowercase, as it is
not a constant. The 'setUp' method was moved above the 'cleanup' method
to respect the flow of time in the Universe.

Change-Id: If0eb17164c8c8bfdee68761aa15ad0eedbacb957