Use both unit and functional for coverage testing
Placement has always relied heavily on functional testing for both API
and database testing, so not including those tests in 'coverage' leads
to misleading results.

This change adjusts the cover job in tox to use a test path of
placement/tests, resulting in all the tests running.

In the process .coveragerc is updated to be placement oriented
and make sure we do not ignore branches.

This is in preparation for turning on a check job for coverage.
See this email for more on that:

http://lists.openstack.org/pipermail/openstack-dev/2018-September/134385.html

A followup patch will turn on a check job for it.

Change-Id: I6c122facc9c9f6e0b2a1d69c6aaf5faeee7cf253