Update code and opts in conf/paths.py for placement
Most of the options in paths.py are not used by placement, so they
are removed. Those that remain are used, mostly by tests, so the
descriptions associated with opts are updated accordingly.

The irrelevance of the removed opts was discovered while doing
the database configuration cleanup that follows in change
Iecf5006b32ca7bc2d46b7a59e78b4941a2e98122 .

Change-Id: I91361ea31c015542be52f905d75d2ca53c61103f