Rationalize and clarify database configuration
Discussion in IRC, hangouts and summarized in email [1] led to a
decision that in placement the use of a default data connection string
was dangerous for people migrating from placement-in-nova to placement
by itself. Instead what we want are explicit but small required actions
for operators/deployers and a hard fail if they do not do them.

A critical first step to make that work is removing the default
connection string and requiring an explicit
[placement_database]/connection value. To whit, this patch:

* removes api_database configuration settings
* removes the default connection value
* make it the value required (which causes config parsing to fail)
* updates test fixtures to remove knowledge of other databases besides
  'placement_database'
* adds a unit tests to confirm cfg.RequiredOptError is raised when
  expected (when the setting is not set)
* Comments in the wsgi app that a hard fail will happen when config
  and required settings are missing

In local testing I have confirmed two different hard failures that we
want:

* If there is no configuration file at all (either in the default
  /etc/placement/placement.conf location or the same file set by
  OS_PLACEMENT_CONFIG_DIR), fail
* If there is a config file and it has no [placement_database]/connection,
  fail.

Note that all these changes are not fully meaningful until the change
in Ibed4c64fa4e74c830d0d2fa040089bf9dc7b8f29, which changes the
placement/wsgi.py app. That's currently the only external entry
point to using placement, and will not work until those changes
are in place.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-September/134444.html

Change-Id: Iecf5006b32ca7bc2d46b7a59e78b4941a2e98122