Add logging_error_fixture to functional tests
The fixture will cause a formatting error when writing logs
to be raised as an exception (thus failing tests).

In the non-gabbi tests the fixture is started in the normal way
with useFixture. In the gabbi tests the fixture setUp and
cleanUp is done manually because the fixtures start and stop
per yaml file, not per test.

Change-Id: Ia5094c3a12bcb9d19d429fbe2164b4533ca9c715