Update the HACKING.rst file
Since the hacking.py file was omitted from the history extraction, it is
felt that it would be best to start the hacking rules from scratch and
build this file (and a new hacking.py) as we need. A NOTE has been added
to clearly mark this file as a WIP for anyone who reads it.

Change-Id: I14cb451a11c680ecca1a43154ecc9d5b9174ab01