Add alloc cands test with nested and aggregates
When placement picks up allocation candidates, the aggregates of
nested providers are assumed as the same as root providers. This
means it ignores the aggregates of the nested provider itself.
This could result in the lack of allocation candidates when an
aggregate on a nested provider but not on the root has been
specified in the `member_of` query parameter.

This patch adds test cases for the bug. The fix is done in a
follow up.

Related-Bug: #1792503

Change-Id: Ic532bc231e89eaec4cb5ad01d27bbe840b01a5f7