Add nova database migration script for postgresql
This is based on the mysql-migrate-db.sh script but
with the necessary changes for interacting with postgresql.

Long-term it might be possible to refactor both the mysql
and postgresql scripts into a generic script but the
potential complexity in doing that makes it a lower priority
than simply getting pg support for the nova_api/placement
table migration done in Stein.

Change-Id: I1cab3183067ab2e41663ca369509d753442e775c