Remove placement.db.migration
This is part of a series of patches to remove noise from the
db migration in placement so that if/when we switch to alembic
there isn't confusing stuff in the way.

This change removes placement.db.migration which is a mapping to
placemement.db.sqlalchemy.migration. Since we only use sqlalchemy
we don't need the mapping.

Change-Id: I9ad49ecf6fc9958d4181b3cbb749185eaacffb9d