Remove support for multiple database from migration.py
This is part of a series of patches to remove noise from the
db migration in placement so that if/when we switch to alembic
there isn't confusing stuff in the way.

In this change the code for database migrations is adjusted to
no longer support multiple databases. While that support was
needed in nova, it is not in placement.

Change-Id: Idcf5f9919fda96b3cd88775ca68a13f20b6eefd2