Placement: Remove usage of get_legacy_facade()
Warnings emitted to the python 3 unit tests:

  OsloDBDeprecationWarning: EngineFacade is deprecated; please use
  oslo_db.sqlalchemy.enginefacade

stem from our use of the get_legacy_facade() oslo.db method.

Following on from [1], this replaces the get_legacy_facade() usage with
the writer transaction context for the placement database. This is done
in both openstack/placement and openstack/nova under the same Change-Id.

[1] If1e355d6174f15a1880e97115a2cf3f9f4be837e

Change-Id: I5f3a7c2c81bd908be30f7bcd9cda7e0ff5b48226