Clean up and clarify tox.ini
While looking into making py36 available when running tox without '-e'
and making py37 more available for both unit testing and functional
testing, it became clear that our tox.ini was a bit messy. This change
cleans it up to:

* Removes all py* jobs and lets tox define those automatically. Their
  commands come from the base [testenv].

  Non-unit tests need to define their commands specifically.

* move pep-related tox below both unit and functional

* remove some out of date comments

* set the default when using tox without -e to unit and functional
  in 2.7 and 3.6, and pep8

* removes the 'stestr slowest' from the unit and functional test env's.
  It's not used consistently and the output is far from consistent so
  it doesn't seem to have any particular value (for this project)

This is one of several ways to do this cleanup. This way assumes some
knowledge of the fact that tox defines a lot default jobs with default
python versions (via factors[1]).

Note that the releasenotes job does not currently work and did
not work before these changes, a conf.py is missing.

[1] https://tox.readthedocs.io/en/latest/config.html#factors-and-factor-conditional-settings

Change-Id: I9a877439e13eb753c7f98dc02e41db891ce33690