Fix genpolicy tox job
The etc/placement/policy-generator.conf file still referred
to nova. Now that it does not, policy generation happens as
expected. The name of the generated file has been changed
to policy.yaml.sample, as we no longer need to distinguish
from nova's.

Note that the readme still points to nova docs, because that
is where we currently have the docs. This will need to change
later.

Change-Id: I633431933666c15eaf8da7e91769d0b4335e6c8d