Make tox -ereleasenotes work
One of the things Placement will need when it is properly publishing
itself is a working release notes job. Until this change there was
no source/conf.py or source/index.rst for releasenotes job to use.
This adds it, we can now create release notes output, with the
openstackdocs theme.

However: Because the placement repo has no branches, all the existing
release notes in releasenotes/notes are considered 'unreleased', which
makes for a lot of noise at the page generated by unreleased.rst.

What this probably means is that we will want to clear out all the
release notes that pre-date extraction and start anew. An experiment to
try removing some notes shows that the releasenotes will produce some
messages about missing release notes, but neither warn nor error, and
the output is good, so this should work out okay for us.

That will be easy to do now that we have a working releasenotes tox env.

Note: This change does not add any jobs to zuul, it just makes tox
work for releasenotes.

Change-Id: I61a7ce492395e0ae69dd0401a141ea4bb93516b9