Move ensure_consumer to a new placement.handlers.util
The driving reason for this change is to allow the placement.db_api
module to be able, in subsequent patches, to import generic utility
routines from placement.util without causing an cyclical import.

It is moved into the handlers package because the only caller is
in there too (in allocation.py).

Relevant unit test are moved, and the callers of ensure_consumer are
updated to reflect the new location.

Change-Id: Ifd1b10e5568b9d6f641e396c3c49101901daff31