Start a contributor goals document
As explained in the document, this is a way to keep track of long term
ideas or hopes that are too far away or too vague to warrant specs, even
wishlist specs. The main goal of these goals is to keep an eye forward
to long term improvements which are not really related to features, but
instead are directed towards making contribution and maintenance more
pleasant.

Change-Id: Ic76852d3bea36f797dd845fd7be520670d214aa3