Allow placement to start without a config file
With oslo.config 6.7.0 configuration options may be found in environment
variables. Because of the small number of options that placement
requires to run it is straightforward to source all config from the
environment, useful in container-based environments.

However, prior to this change, placement-the-wsgi-app required that a
configuration file existed, placement.conf, either in /etc/placement or
in a custom directory.

This change makes it so that the placement.conf can be any of:

* in the custom directory
* in the default project locations (including /etc/placement)
* nowhere

The change maintains the behavior that if
[placement_database]/connection is not set the application will fail to
start.

Though this change is orthogonal to the oslo.config change, requirements
and lower-constraints are updated to ensure the desired behavior.

Closes-Bug: #1802925
Change-Id: Iefa8ad22dcb6a128293ea71ab77c377db56e8d70