Documentation cleanup: front page
This is the first in a series of changes to the content in
doc/source to make it more accurately reflect the current state of
placement, now that it is in its own repo.

There are small changes throughout the front page, mostly to make
it more clear and make useful stuff a bit more discoverable.

The two major changes are:

* Older (pre-stein) upgrade notes are found by linking back to
  nova's rocky documentation.
* The REST API history is moved to its own sub-page, otherwise it
  the page is overwhelming and noisy.

Close review is probably required. cdent typed this. That has
implications.

Change-Id: I66e0c7d18b253b0a5a8fdac65e30b5b3cef37db2