Remove keystoneauth1 opts from placement config group
The keystoneauth1 opts were included in the [placement] config group
in nova to define how nova was supposed to talk to the placement
service that it discovered in its service catalog.

Placement doesn't need to talk to itself, and for the time being it
doesn't need to talk to anyone else but the database, so we can
remove the settings. We do not need to deprecate them because we
have already declared that a "fresh config" is required and we
haven't had a release yet. If people have the settings in their
config files they simply won't be used.

Similar settings are registered in the [keystone] group which
will be cleaned up in a separate change (note that the [keystone]
group is not the same as the [keystone_authtoken] group which
is used by the authtoken API middleware).

There's no test coverage for these because any tests which would
have used them are still in nova. Integration tests will make
sure this change is fine.

Change-Id: I180a4df89d8338dc640dc2c5638775c65bde1cb7