Add assertions to verify placement-manage error output
The tests for placement-manage, in unit/cmd/test_manage were not
verifying help message output when partial or incorrect commands
were provided. This led to some confusion when trying to verify
a fix related to different behaviors of argparse.

Here, some verification is added. Because of differences between
python 2 and 3, we need to look at stderr and stdout, respectively, to
verify the output.

Change-Id: I337a25cfdc23bc56fbf15f128a94d2b1b1d1b664
Related-Bug: #1804420