Remove [keystone] config options from placement
Similar to I180a4df89d8338dc640dc2c5638775c65bde1cb7, this
removes the entire [keystone] group from placement.conf
since placement does not use it. The [keystone] group from
nova.conf was optionally used by a couple of compute APIs for
validating a user-provided project ID (flavor access and quota APIs).
Since placement doesn't use the options, they are removed here
so we don't have to go through a formal deprecation period after
placement is officially released. The placement/conf/utils.py
module is also not used so it is removed as well.

While doing this, it was noticed that the [keystone_authtoken]
options, which *are* used by placement.auth.PlacementAuthProtocol
for the auth token middleware, were missing from the generated
placement config docs, so that's fixed in this change as well.

Change-Id: Ic856a4a43a21fbae3fe98466d1dc716d2736d622